requirements.txt
torch>=2.0
torchvision>=0.15
sam2 @ git+https://github.com/facebookresearch/sam2.git
onnx>=1.15
onnxruntime-gpu>=1.17
tensorrt>=10.0
cuda-python
numpy>=1.24
opencv-python>=4.8
Pillow>=10.0
pyyaml>=6.0
tqdm
